Checkpoint
Reworked the code that handles fsevent notifications. A posix thread is
created when the start method is called. The fsevent run loop is started
within this thread so as to avoid hanging the ruby interpreter.

When a file system notification occurs, the directory paths are sent into a
pipe as a single string with a newline between each path.

A method is provided to read out these changed directories as a ruby array of
strings.