Reading all notifications from the pipe.
When the "changes" method is invoked on the notifier, all messages will be
read from the pipe and returned to the user. The returned array might contain
duplicates - the user can call uniq if they need it.

Also using the rb_thread_select method -- plays more nicely with the ruby
thread scheduler than the simpleminded call to a raw select.